Implement fetch_source
FetchSource takes in a source object with at least a url attributes, and
persists it to the data store. This is done by opening the url, passing
the data to RSS::Parser and returning the results to the interactor.
Once the interactor has the results it iterates over the items, maps
each one to a new Struct and then persists them to the persistance layer